requirements.txt
transformers 
Pillow 
opencv-python 
torch
gtts
pygame